move db open/close functions to database crate
