+++ simpa/Cargo.toml
@@ -21,6 +21,7 @@ clap.workspace = true
 indexmap.workspace = true
 log.workspace = true
 async-channel.workspace = true
+num_cpus.workspace = true
 
 futures = "0.3"
 rand_distr = "0.4"
